[BE] refactor: group dto 리팩토링
