--- server/src/domains/group/dto/ApplyedGroupQuery.ts
@@ -1,7 +1,0 @@
-import { IsEnum } from 'class-validator';
-import { ApplyGroupState } from '../models/ApplyGroup';
-
-export class ApplyedGroupQuery {
-  @IsEnum(ApplyGroupState)
-  state: ApplyGroupState;
-}
