+++ server/src/domains/group/dto/query/ApplyedGroupQuery.ts
@@ -1,0 +1,7 @@
+import { ApplyGroupState } from '@domains/group/models/ApplyGroup';
+import { IsEnum } from 'class-validator';
+
+export class ApplyedGroupQuery {
+  @IsEnum(ApplyGroupState)
+  state: ApplyGroupState;
+}
