+++ server/src/domains/group/dto/query/GroupQuery.ts
@@ -1,0 +1,19 @@
+import { IsNumber, IsOptional, IsString } from 'class-validator';
+
+export class GroupQuery {
+  @IsOptional()
+  @IsNumber()
+  page: number;
+
+  @IsOptional()
+  @IsString()
+  name: string;
+
+  @IsOptional()
+  @IsNumber()
+  category: number;
+
+  @IsOptional()
+  @IsString()
+  techstack: string;
+}
