+++ server/src/domains/group/dto/request/ApplyGroupDto.ts
@@ -1,4 +1,4 @@
-import { IsNumber, IsString } from 'class-validator';
+import { IsNumber } from 'class-validator';
 
 export class ApplyGroupDto {
   @IsNumber()
