+++ server/src/domains/group/dto/response/GroupRoleResponse.ts
@@ -1,5 +1,5 @@
 import { IsEnum } from 'class-validator';
-import { GroupEnrollmentAs } from '../models/GroupEnrollment';
+import { GroupEnrollmentAs } from '../../models/GroupEnrollment';
 
 export class GroupRoleResponse {
   @IsEnum(GroupEnrollmentAs)
